// clients/js-kit/src/generated/types/baseExtraAccount.ts
export type BaseExtraAccount =
  | { __kind: 'PreconfiguredProgram'; isSigner: boolean; isWritable: boolean }
  | {
      __kind: 'PreconfiguredCollection';
      isSigner: boolean;
      isWritable: boolean;
    }
  | { __kind: 'PreconfiguredOwner'; isSigner: boolean; isWritable: boolean }
  | { __kind: 'PreconfiguredRecipient'; isSigner: boolean; isWritable: boolean }
  | { __kind: 'PreconfiguredAsset'; isSigner: boolean; isWritable: boolean }
  | {
      __kind: 'CustomPda';
      seeds: Array<BaseSeed>;
      customProgramId: Option<Address>;
      isSigner: boolean;
      isWritable: boolean;
    }
  | {
      __kind: 'Address';
      address: Address;
      isSigner: boolean;
      isWritable: boolean;
    };